requirements_dev.txt
influxdb_client==1.36.1
iperf3==0.1.11
python-dotenv==1.0.0
netifaces-plus==0.12.0
pytest==7.3.2
numpy==1.24.3